+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `field_tolerance` field/s to look for resistor tolerance.
   - `default_resistor_tolerance` which tolerance to use when none found.
   - `cache_3d_resistors` to avoid generating them all the time.
+  - `resources_dir` to specify fonts and colors to install (CI/CD)
 - 3D: colored 3D models for THT resistors
 - Datasheet download: now the warnings mention which reference failed.
 - Plot related outputs and PCB_Print: